+++ src/string_util.h
@@ -25,6 +25,7 @@
  */
 #include <string.h>
 #include <syslog.h>
+#include <sys/types.h>
 
 #include <string>
 #include <sstream>
@@ -47,7 +48,6 @@ std::string lower(std::string s);
 std::string IntToStr(int);
 std::string get_date();
 std::string urlEncode(const std::string &s);
-std::string prepare_url(const char* url);
 bool get_keyword_value(std::string& target, const char* keyword, std::string& value);
 
 #endif // S3FS_STRING_UTIL_H_
